requirements.txt
azure.ai.documentintelligence
azure-identity
azure.storage.blob
python-dotenv